+++ Project/src/methods/hybrid/xder.py
@@ -71,12 +71,13 @@ class XDER(BaseCLMethod):
                         < mem["logit_sizes"][:, None]
                     )
                     dark_loss = F.mse_loss(
-                        buf_logits_now[:, :stored_dim][mask],
-                        mem["logits"][mask],
+                        buf_logits_now[:, :stored_dim].float()[mask],
+                        mem["logits"].float()[mask],
                     )
                 else:
                     dark_loss = F.mse_loss(
-                        buf_logits_now[:, :stored_dim], mem["logits"]
+                        buf_logits_now[:, :stored_dim].float(),
+                        mem["logits"].float(),
                     )
                 # Hard-label calibration.
                 cal_loss = F.cross_entropy(buf_logits_now, mem["y"])
@@ -91,7 +92,7 @@ class XDER(BaseCLMethod):
         with torch.no_grad():
             with torch.cuda.amp.autocast(enabled=self.fp16):
                 stored_logits = self.model(x)
-        self.buffer.add(x, y, task_id, logits=stored_logits.detach())
+        self.buffer.add(x, y, task_id, logits=stored_logits.detach().float())
 
         return loss.item()
